[refactor|test] Extract generators/properties for targets
