+++ tests/include/generators/types/callable-type.hpp
@@ -8,7 +8,7 @@
 
 #include <type_traits>
 
-#include "generators/composable-callable.hpp"
+#include "generators/values/composable-callable.hpp"
 
 namespace Callable
 {
