+++ tests/include/generators/values/runnable-combination.hpp
@@ -6,9 +6,9 @@
 
 #pragma once
 
-#include "generators/composable-callable.hpp"
-#include "generators/foldable-callable.hpp"
-#include "generators/productable-container.hpp"
+#include "generators/values/composable-callable.hpp"
+#include "generators/values/foldable-callable.hpp"
+#include "generators/values/productable-container.hpp"
 
 namespace Productable
 {
