+++ tests/include/generators/values/runnable-double-combination.hpp
@@ -6,7 +6,7 @@
 
 #pragma once
 
-#include "generators/runnable-combination.hpp"
+#include "generators/values/runnable-combination.hpp"
 
 namespace Runnable
 {
