+++ tests/src/instanciation/productable-container.cpp
@@ -4,7 +4,7 @@
  * SPDX - License - Identifier: MIT
  */
 
-#include "generators/productable-container.hpp"
+#include "generators/values/productable-container.hpp"
 
 template class Productable::ProductableContainer<0, decltype(std::make_index_sequence<1>()), 0>;
 template class Productable::ProductableContainer<0, decltype(std::make_index_sequence<2>()), 0>;
